+++ src/Data/Functor/PushPull.hs
@@ -14,6 +14,7 @@ module Data.Functor.PushPull where
 import Data.Typeable
 import Control.Elevator
 import Data.OpenUnion1.Clean
+import Control.Applicative
 
 -- | The type for asynchronous input/output.
 data PushPull a b r = Push a r | Pull (b -> r) deriving (Functor, Typeable)
@@ -27,3 +28,10 @@ push a = elevate (Push a ())
 
 pull :: (Elevate (PushPull a b) f) => f b
 pull = elevate (Pull id)
+
+-- | @filterPush :: (a -> Bool) -> PushPull a b r -> Program (PushPull a b) r@
+filterPush :: (Applicative f, Elevate (PushPull a b) f) => (a -> Bool) -> PushPull a b r -> f r
+filterPush p e@(Push a r)
+  | p a = elevate e
+  | otherwise = pure r
+filterPush _ e = elevate e